feat: Implement "scroll-back" to latest message when navigating replies
This commit adds a navigation history for quoted messages, such as if
Alice navigated to message A by pressing the "quoted message" inside the
reply message B and pressed the "scroll down button", the chat will be
scrolled back to message B instead of the very bottom. This is true for
as many replies as the user navigates through.
All the messages higher than the lowest fully visible message
are automatically cleared from the stack as the user scrolls.

This feature makes it easier to read threaded conversations,
especially in scenarios with much to catch up on. This is most useful in
group chats and communities.